feat(settings): add update part
