+++ src/browsers/index.js
@@ -5,7 +5,6 @@ module.exports = (dirname, storage, util) => ({
   colorsbook: require('./colorsbook')(dirname, storage, util),
   picker: require('./picker')(dirname, storage, util),
   preview: require('./preview')(dirname, storage, util),
-  about: require('./about')(dirname, storage, util),
   settings: require('./settings')(dirname, storage, util),
   support: require('./support')(dirname, storage, util)
 })
