+++ src/events/index.js
@@ -5,6 +5,5 @@ module.exports = (storage, browsers, eventEmitter) => ({
   colorsbook: require('./colorsbook')(storage, browsers, eventEmitter),
   picker: require('./picker')(storage, browsers, eventEmitter),
   preview: require('./preview')(storage, browsers, eventEmitter),
-  about: require('./about')(storage, browsers, eventEmitter),
   settings: require('./settings')(storage, browsers, eventEmitter)
 })
