--- src/views/about.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>About Colorpicker</title>
-    <link rel="stylesheet" href="./css/about.css" />
-    <link rel="stylesheet" href="./css/font-awesome.min.css" />
-    <script defer src="./js/about.js"></script>
-  </head>
-  <body>
-    <img src="../../build/icon.png" alt="Logo Colorpicker" />
-    <h1>Colorpicker</h1>
-    <section>
-      <p id="version">Release <span>{...}</span></p>
-      <p id="update"><i class="fa fa-refresh fa-spin fa-3x fa-fw"></i> Checking for new update...</p>
-    </section>
-    <p id="thanks">Thank's to Ryunako, Hanabi &amp; Khaly!</p>
-  </body>
-</html>
