+++ src/views/settings.html
@@ -92,8 +92,9 @@
       <p>Chrome - version <span id="chrome-version"></span></p>
       <p>V8 - version <span id="v8-version"></span></p>
       <h2>Update</h2>
-      <p>No update available</p>
+      <p id="update">Fetching..</p>
       <h2>Thank's part</h2>
+      <p>Thank's to Ryunako, Khaly, PAN, and Hanabi!</p>
       <p>Colorpicker is developed by Toinane with <i class="fa fa-heart"></i></p>
     </section>
   </body>
